# Replace Span with a protobuf model