--- a/model/top_level_test.go
+++ b/model/top_level_test.go
@@ -20,15 +20,10 @@ func TestTopLevelTypical(t *testing.T) {
 	tr.ComputeTopLevel()
 
 	assert.True(tr[0].TopLevel(), "root span should be top-level")
-	assert.True(tr[0].topLevel, "root span should be top-level")
 	assert.False(tr[1].TopLevel(), "main service, and not a root span, not top-level")
-	assert.False(tr[1].topLevel, "main service, and not a root span, not top-level")
 	assert.True(tr[2].TopLevel(), "only 1 span for this service, should be top-level")
-	assert.True(tr[2].topLevel, "only 1 span for this service, should be top-level")
 	assert.True(tr[3].TopLevel(), "only 1 span for this service, should be top-level")
-	assert.True(tr[3].topLevel, "only 1 span for this service, should be top-level")
 	assert.False(tr[4].TopLevel(), "yet another sup span, not top-level")
-	assert.False(tr[4].topLevel, "yet another sup span, not top-level")
 }
 
 func TestTopLevelSingle(t *testing.T) {
@@ -67,15 +62,10 @@ func TestTopLevelOneService(t *testing.T) {
 	tr.ComputeTopLevel()
 
 	assert.False(tr[0].TopLevel(), "just a sub-span, not top-level")
-	assert.False(tr[0].topLevel, "just a sub-span, not top-level")
 	assert.False(tr[1].TopLevel(), "just a sub-span, not top-level")
-	assert.False(tr[1].topLevel, "just a sub-span, not top-level")
 	assert.True(tr[2].TopLevel(), "root span should be top-level")
-	assert.True(tr[2].topLevel, "root span should be top-level")
 	assert.False(tr[3].TopLevel(), "just a sub-span, not top-level")
-	assert.False(tr[3].topLevel, "just a sub-span, not top-level")
 	assert.False(tr[4].TopLevel(), "just a sub-span, not top-level")
-	assert.False(tr[4].topLevel, "just a sub-span, not top-level")
 }
 
 func TestTopLevelLocalRoot(t *testing.T) {
@@ -94,19 +84,12 @@ func TestTopLevelLocalRoot(t *testing.T) {
 	tr.ComputeTopLevel()
 
 	assert.True(tr[0].TopLevel(), "root span should be top-level")
-	assert.True(tr[0].topLevel, "root span should be top-level")
 	assert.False(tr[1].TopLevel(), "main service, and not a root span, not top-level")
-	assert.False(tr[1].topLevel, "main service, and not a root span, not top-level")
 	assert.True(tr[2].TopLevel(), "only 1 span for this service, should be top-level")
-	assert.True(tr[2].topLevel, "only 1 span for this service, should be top-level")
 	assert.True(tr[3].TopLevel(), "top-level but not root")
-	assert.True(tr[3].topLevel, "top-level but not root")
 	assert.False(tr[4].TopLevel(), "yet another sup span, not top-level")
-	assert.False(tr[4].topLevel, "yet another sup span, not top-level")
 	assert.False(tr[5].TopLevel(), "yet another sup span, not top-level")
-	assert.False(tr[5].topLevel, "yet another sup span, not top-level")
 	assert.False(tr[6].TopLevel(), "yet another sup span, not top-level")
-	assert.False(tr[6].topLevel, "yet another sup span, not top-level")
 }
 
 func TestTopLevelWithTag(t *testing.T) {
@@ -122,10 +105,8 @@ func TestTopLevelWithTag(t *testing.T) {
 	t.Logf("%v\n", tr[1].Metrics)
 
 	assert.True(tr[0].TopLevel(), "root span should be top-level")
-	assert.True(tr[0].topLevel, "root span should be top-level")
 	assert.Equal(float64(42), tr[0].Metrics["custom"], "custom metric should still be here")
 	assert.False(tr[1].TopLevel(), "not a top-level span")
-	assert.False(tr[1].topLevel, "not a top-level span")
 	assert.Equal(float64(42), tr[1].Metrics["custom"], "custom metric should still be here")
 }
 
@@ -137,20 +118,16 @@ func TestTopLevelGetSetBlackBox(t *testing.T) {
 	assert.False(span.TopLevel(), "by default, all spans are considered non top-level")
 	span.setTopLevel(true)
 	assert.True(span.TopLevel(), "marked as top-level")
-	assert.True(span.topLevel, "marked as top-level")
 	span.setTopLevel(false)
 	assert.False(span.TopLevel(), "no more top-level")
-	assert.False(span.topLevel, "no more top-level")
 
 	span.Metrics = map[string]float64{"custom": 42}
 
 	assert.False(span.TopLevel(), "by default, all spans are considered non top-level")
 	span.setTopLevel(true)
 	assert.True(span.TopLevel(), "marked as top-level")
-	assert.True(span.topLevel, "marked as top-level")
 	span.setTopLevel(false)
 	assert.False(span.TopLevel(), "no more top-level")
-	assert.False(span.topLevel, "no more top-level")
 }
 
 func TestTopLevelGetSetMetrics(t *testing.T) {