… model/span.pb.go
… model/trace_test.go